fix: move models to their own modules - DIA-61984
