+++ fastapi_sqla/__init__.py
@@ -3,11 +3,9 @@ import os
 from fastapi import FastAPI
 
 from fastapi_sqla import sqla
+from fastapi_sqla.models import Collection, Item, Page
 from fastapi_sqla.sqla import (
     Base,
-    Collection,
-    Item,
-    Page,
     Paginate,
     PaginateSignature,
     Pagination,
